refactor(core): extract shared config-object helpers
